requirements.txt
flask
flask-cors
pypdf
langchain_community
langchain-openai
langchain
python-dotenv
